Change pending request gauge to incoming request counter
Pending count is not accurate enough, somehow all increments for it
are not matched by decrements, and over time it keeps increasing and
shows hundreds of pending requests, even when there are none.

As this could not be fixed with current OPEA code structure, changed
it to a total incoming requests counter. That can be post-processed to
get incoming requests rate.

However, that reflects more service perf than its request load, it's
not really good for scaling rest of application services to increase
the perf, like pending reqs count (= queue size) would be.

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>